seemingly working setup for dev and production.
- Had to shuffle around some directory references, /dist/, etc.
- Reintroduced Owen's catchall route for index.html
  - Should maybe be explicit about which routes are supposed to redirect to /index
  - ... "isomorphic" router so we don't define the same logic in two places?
- Little peachworks API proxy module factored out so server is less messy
