+++ index.html
@@ -56,7 +56,6 @@
   <body>
     <div id='root'>
     </div>
-    <!-- adding dist makes dev server work, but breaks netlify, but we're not using netlify anymore..? -->
-    <script src="bundle.js"></script>
+    <script src="/dist/bundle.js"></script>
   </body>
 </html>
