+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "clean": "rimraf dist",
     "build:webpack": "NODE_ENV=production webpack --config webpack.config.prod.js",
-    "build": "npm run clean && npm run build:webpack && sh setup_dist.sh",
+    "build": "npm run clean && npm run build:webpack",
     "lint": "eslint src",
     "start": "node server/index.js",
     "test": "mocha --compilers js:babel-register --recursive",
