--- setup_dist.sh
@@ -1,2 +1,0 @@
-echo "copying index into dist"
-cp index.html dist
