+++ webpack.config.prod.js
@@ -15,8 +15,7 @@ module.exports = {
     new webpack.optimize.OccurenceOrderPlugin(),
     new webpack.DefinePlugin({
       'process.env': {
-        'NODE_ENV': JSON.stringify('production'),
-        'PEACHWORKS_ACCESS_TOKEN': JSON.stringify(process.env.PEACHWORKS_ACCESS_TOKEN)
+        'NODE_ENV': JSON.stringify('production')
       }
     }),
     new webpack.optimize.UglifyJsPlugin({
